backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
aiohttp==3.11.11
beautifulsoup4==4.12.3
google-genai==1.14.0
google-search-results==2.4.2
pydantic==2.10.4
jinja2==3.1.5
xhtml2pdf==0.2.17
python-dotenv==1.0.1
platformdirs==4.3.6
pyinstaller==6.19.0
